backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
sqlalchemy>=2.0.25
alembic>=1.13.1
psycopg2-binary>=2.9.9
pydantic>=2.5.3
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
pytest>=8.2
hypothesis>=6.98.3
bcrypt>=4.1.2
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
httpx>=0.25.0,<0.28
Pillow>=10.0.0
slowapi>=0.1.9
aiosmtplib>=2.0.0
twilio>=8.0.0
openai>=1.0.0
email-validator>=2.0.0
piexif>=1.1.3
